+++ Services/Interfaces/InterfacesLab4/IOrdinarySortingService4.cs
@@ -2,7 +2,7 @@
 
 namespace Laba.Services.Interfaces.InterfacesLab4
 {
-    public interface IOrdinarySortingService4 : ISorting<int[], int>, ISwap<int>
+    public interface IOrdinarySortingService4 : ISorting<int[], double>, ISwap<int>
     {
         void Merge(int[] array, int start, int end, int mid);
         void RecursiveSort(int[] array, int start, int end);
