+++ Services/Interfaces/InterfacesLab5/IOrdinarySortingService5.cs
@@ -2,7 +2,7 @@
 
 namespace Laba.Services.Interfaces.InterfacesLab5
 {
-    public interface IOrdinarySortingService5 : ISorting<int[], int>, ISwap<int>
+    public interface IOrdinarySortingService5 : ISorting<int[], double>, ISwap<int>
     {
         public void PopulateBounds(int lower, int upper, int length);
     }
